+++ NetKnife/modules/__init__.py
@@ -1,0 +1,13 @@
+from scapy.all import *
+from fakeap import *
+from arp import *
+from ssidhidden import *
+from hashgen import *
+from syn import *
+from netsniffer import *
+from apd import *
+from ipchange import *
+from deauthenticatio import *
+from tcpport import *
+from netscan import *
+from macflood import *
