+++ NetKnife/modules/fakeap.py
@@ -1,0 +1,28 @@
+from scapy import *
+from colorama import Fore
+#[2]:Fake Access Point Creator Tools
+#This tool is used to create fake access points
+
+def FakeAP():
+
+    print """
+    .########.##....##....###....########.
+    .##.......##...##....##.##...##.....##
+    .##.......##..##....##...##..##.....##
+    .######...#####....##.....##.########.
+    .##.......##..##...#########.##.......
+    .##.......##...##..##.....##.##.......
+    .##.......##....##.##.....##.##.......
+    """
+    iface = raw_input("Plese Enter The Interface (WireLess Net Adapter On Monitor Mode): ")
+    sender_mac = RandMAC()
+    ssid = raw_input("Plese Enter The SSID(Name Of AP):" )
+    dot11 = Dot11(type=0, subtype=8, addr1="ff:ff:ff:ff:ff:ff", addr2=sender_mac, addr3=sender_mac)
+    beacon = Dot11Beacon()
+    essid = Dot11Elt(ID="SSID", info=ssid, len=len(ssid))
+    frame = RadioTap()/dot11/beacon/essid
+    print Fore.RED + "Created AP"
+    print "AP Name: " + ssid
+    print "interfaces: " + iface
+
+    sendp(frame, inter=0.1, iface=iface, loop=1)
