+++ NetKnife/modules/netsniffer.py
@@ -1,0 +1,25 @@
+
+#[6]:Network Packet Capture Tools
+#This tool is used to record ongoing traffic from the interface
+from colorama import Fore
+from scapy.all import *
+def NetSniffer():
+
+
+    print """
+    .##....##.########.########..######..##....##.####.########.########.########.########.
+    .###...##.##..........##....##....##.###...##..##..##.......##.......##.......##.....##
+    .####..##.##..........##....##.......####..##..##..##.......##.......##.......##.....##
+    .##.##.##.######......##.....######..##.##.##..##..######...######...######...########.
+    .##..####.##..........##..........##.##..####..##..##.......##.......##.......##...##..
+    .##...###.##..........##....##....##.##...###..##..##.......##.......##.......##....##.
+    .##....##.########....##.....######..##....##.####.##.......##.......########.##.....##
+    """
+    iface = raw_input("Plese Enter The Interface For Sniffing: ")
+    def PacketSniifer(packet):
+       print packet.summary()
+       return
+
+    sniff(iface=iface, count="100", prn=PacketSniifer)
+
+
